Add files via upload
The IR sensor is strategically placed in a location where it can detect the movement of individuals within its sensing range.The IR sensor continuously monitors its surroundings for any changes in infrared radiation caused by the movement of individuals.When someone enters the detection range of the IR sensor, it detects the change in radiation and triggers an output signal.When the IR sensor detects motion, it sends a signal to the microcontroller which then activates the LED lighting system. The LED lights up.
Automatic light systems contribute to energy conservation by ensuring that lights are not left on unnecessarily when the area is unoccupied.These systems can improve accessibility for individuals with disabilities by providing automatic illumination in response to their movement.
The automatic staircase lighting system is a lighting control system where the lights are automated according to the presence of people using the staircase. It automatically switches ON lights when the sunlight goes below the visible region of our eyes and when person is present. This is done by a sensor called Light Dependant Resistor (LDR) which senses the light actually like our eyes and IR sensor. It automatically switches OFF lights whenever the sunlight comes, visible to our eyes.